requirements.txt
fastapi
uvicorn
python-dotenv
cohere
pinecone-client
anthropic